# web-morda/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Cookbook",
    "version": "1.0"
  },
  "paths": {
    "/api/dishes": {
      "post": {
        "tags": [
          "Dishes"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Photos": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "PhotoFiles": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "Calories": {
                    "type": "number",
                    "format": "float"
                  },
                  "Proteins": {
                    "type": "number",
                    "format": "float"
                  },
                  "Fats": {
                    "type": "number",
                    "format": "float"
                  },
                  "Carbs": {
                    "type": "number",
                    "format": "float"
                  },
                  "Products": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DishProductRequest"
                    }
                  },
                  "PortionSize": {
                    "type": "number",
                    "format": "float"
                  },
                  "Category": {
                    "$ref": "#/components/schemas/DishCategory"
                  },
                  "Flags": {
                    "uniqueItems": true,
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DietaryFlag"
                    }
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Photos": {
                  "style": "form"
                },
                "PhotoFiles": {
                  "style": "form"
                },
                "Calories": {
                  "style": "form"
                },
                "Proteins": {
                  "style": "form"
                },
                "Fats": {
                  "style": "form"
                },
                "Carbs": {
                  "style": "form"
                },
                "Products": {
                  "style": "form"
                },
                "PortionSize": {
                  "style": "form"
                },
                "Category": {
                  "style": "form"
                },
                "Flags": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Dishes"
        ],
        "parameters": [
          {
            "name": "Category",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DishCategory"
            }
          },
          {
            "name": "Vegan",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/FlagFilterOption"
            }
          },
          {
            "name": "GlutenFree",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/FlagFilterOption"
            }
          },
          {
            "name": "SugarFree",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/FlagFilterOption"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DishResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DishResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DishResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/dishes/{id}": {
      "get": {
        "tags": [
          "Dishes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Dishes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Photos": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "PhotoFiles": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "Calories": {
                    "type": "number",
                    "format": "float"
                  },
                  "Proteins": {
                    "type": "number",
                    "format": "float"
                  },
                  "Fats": {
                    "type": "number",
                    "format": "float"
                  },
                  "Carbs": {
                    "type": "number",
                    "format": "float"
                  },
                  "Products": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DishProductRequest"
                    }
                  },
                  "PortionSize": {
                    "type": "number",
                    "format": "float"
                  },
                  "Category": {
                    "$ref": "#/components/schemas/DishCategory"
                  },
                  "Flags": {
                    "uniqueItems": true,
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DietaryFlag"
                    }
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Photos": {
                  "style": "form"
                },
                "PhotoFiles": {
                  "style": "form"
                },
                "Calories": {
                  "style": "form"
                },
                "Proteins": {
                  "style": "form"
                },
                "Fats": {
                  "style": "form"
                },
                "Carbs": {
                  "style": "form"
                },
                "Products": {
                  "style": "form"
                },
                "PortionSize": {
                  "style": "form"
                },
                "Category": {
                  "style": "form"
                },
                "Flags": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Dishes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DietaryFlag": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "DishCategory": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7
        ],
        "type": "integer",
        "format": "int32"
      },
      "DishProductRequest": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "DishProductResponse": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "productName": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "DishResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "photos": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "calories": {
            "type": "number",
            "format": "float"
          },
          "proteins": {
            "type": "number",
            "format": "float"
          },
          "fats": {
            "type": "number",
            "format": "float"
          },
          "carbs": {
            "type": "number",
            "format": "float"
          },
          "autoCalculatedNutrition": {
            "$ref": "#/components/schemas/NutritionResponse"
          },
          "portionSize": {
            "type": "number",
            "format": "float"
          },
          "category": {
            "$ref": "#/components/schemas/DishCategory"
          },
          "flags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DietaryFlag"
            },
            "nullable": true
          },
          "availableFlags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DietaryFlag"
            },
            "nullable": true
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DishProductResponse"
            },
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FlagFilterOption": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "NutritionResponse": {
        "type": "object",
        "properties": {
          "calories": {
            "type": "number",
            "format": "float"
          },
          "proteins": {
            "type": "number",
            "format": "float"
          },
          "fats": {
            "type": "number",
            "format": "float"
          },
          "carbs": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}